+++ wallride-core/src/main/java/org/wallride/repository/TagRepositoryImpl.java
@@ -95,7 +95,7 @@ public class TagRepositoryImpl implements TagRepositoryCustom {
 				.createFullTextQuery(searchQuery, Tag.class)
 				.setCriteriaQuery(criteria)
 				.setSort(sort);
-		persistenceQuery.setFirstResult(pageable.getOffset());
+		persistenceQuery.setFirstResult((int) pageable.getOffset());
 		persistenceQuery.setMaxResults(pageable.getPageSize());
 
 		int resultSize = persistenceQuery.getResultSize();
